requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.9.1
factory-boy==2.11.1
Faker==1.0.2
gunicorn==19.9.0
more-itertools==6.0.0
pluggy==0.9.0
psycopg2==2.7.7
py==1.8.0
pytest==4.3.0
pytest-django==3.4.8
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
text-unidecode==1.2
